+++ android/src/main/java/ly/count/dart/countly_flutter/CountlyFlutterPlugin.java
@@ -21,6 +21,7 @@ import ly.count.android.sdk.ModuleFeedback.*;
 import ly.count.android.sdk.DeviceIdType;
 import ly.count.android.sdk.ContentCallback;
 import ly.count.android.sdk.ContentStatus;
+import ly.count.android.sdk.WebViewDisplayOption;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -72,6 +73,7 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
     private final String COUNTLY_FLUTTER_SDK_NAME_NO_PUSH = "dart-flutterbnp-android";
 
     private final boolean BUILDING_WITH_PUSH_DISABLED = false;
+    private static final int DATA_SCHEMA_VERSIONS = 4;
 
     public void notifyPublicChannelRCDL(RequestResult downloadResult, String error, boolean fullValueUpdate, Map<String, RCData> downloadedValues, Integer requestID) {
         Map<String, Object> data = new HashMap<>();
@@ -1415,6 +1417,26 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
                 CountlyStore countlyStore = new CountlyStore(context, new ModuleLog());
                 countlyStore.addRequest(args.getString(0), true);
                 result.success("storeRequest: success");
+            } else if ("setServerConfig".equals(call.method)) {
+                CountlyStore countlyStore = new CountlyStore(context, new ModuleLog());
+                JSONObject jsonObject = args.getJSONObject(0);
+                countlyStore.setServerConfig(jsonObject.toString());
+                // Why this added here, it is that because when it is set something in storage
+                // sdk assumes that it is an older version so it start migrations that needs to be done
+                // but in this case we only want to use setting server config and do not want migrations
+                // to mess up our process flow. So in here we are setting it to latest known to get away with it.
+                // Normally in a fresh install migrations are first to run and they run once.
+                countlyStore.setDataSchemaVersion(DATA_SCHEMA_VERSIONS);
+                result.success("setServerConfig: success");
+            } else if ("getServerConfig".equals(call.method)) {
+                CountlyStore countlyStore = new CountlyStore(context, new ModuleLog());
+                String sc = countlyStore.getServerConfig();
+                Map<String, Object> serverConfigMap = new HashMap<>();
+                try {
+                    serverConfigMap = toMap(new JSONObject(sc));
+                } catch (JSONException ignored) {
+                }
+                result.success(serverConfigMap);
             } else if ("addDirectRequest".equals(call.method)) {
                 JSONObject jsonObject = args.getJSONObject(0);
                 Map<String, String> requestMap = new HashMap<>();
@@ -1427,7 +1449,10 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
             } else if ("halt".equals(call.method)) {
                 Countly.sharedInstance().halt();
                 result.success("halt: success");
-            } else if ("enterContentZone".equals(call.method)) {
+            }
+            //------------------End------------------------------------
+
+            else if ("enterContentZone".equals(call.method)) {
                 Countly.sharedInstance().contents().enterContentZone();
                 result.success(null);
             } else if ("exitContentZone".equals(call.method)) {
@@ -1441,7 +1466,6 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
                 Countly.sharedInstance().contents().previewContent(contentId);
                 result.success(null);
             }
-            //------------------End------------------------------------
 
             else {
                 result.notImplemented();
@@ -1730,6 +1754,10 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
             this.config.setRequestDropAgeHours(_config.getInt("requestDropAgeHours"));
         }
 
+        if (_config.has("requestTimeoutDuration")) {
+            this.config.setRequestTimeoutDuration(_config.getInt("requestTimeoutDuration"));
+        }
+
         if (_config.has("manualSessionEnabled") && _config.getBoolean("manualSessionEnabled")) {
             enableManualSessionControl();
         }
@@ -1795,6 +1823,10 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
             this.config.setGlobalViewSegmentation(toMap(globalViewSegmentation));
         }
 
+        if (_config.has("disableViewRestartForManualRecording") && _config.getBoolean("disableViewRestartForManualRecording")) {
+            this.config.disableViewRestartForManualRecording();
+        }
+
         if (_config.has("enableAllConsents") && _config.getBoolean("enableAllConsents")) {
              this.config.giveAllConsents();
         }
@@ -1813,6 +1845,15 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
 
         if (_config.has("zoneTimerInterval")) {
             this.config.content.setZoneTimerInterval(_config.getInt("zoneTimerInterval"));
+        }
+
+        if (_config.has("webviewDisplayOption")) {
+            String option = _config.getString("webviewDisplayOption");
+            if ("IMMERSIVE".equals(option)) {
+                this.config.setWebviewDisplayOption(WebViewDisplayOption.IMMERSIVE);
+            } else if ("SAFE_AREA".equals(option)) {
+                this.config.setWebviewDisplayOption(WebViewDisplayOption.SAFE_AREA);
+            }
         }
 
         this.config.content.setGlobalContentCallback(new ContentCallback() {