+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## XX.XX.XX
 * Added Content feature method `previewContent(String contentId)` (Experimental!).
+* Added a new config option disableViewRestartForManualRecording to disable auto close/restart behavior of manual views on app background/foreground actions.
+* Added a new config option "setWebviewDisplayOption: WebViewDisplayOption" to control how Content and Feedback Widgets are presented.
+  * IMMERSIVE mode (default): Full-screen display (except cutouts).
+  * SAFE_AREA mode: Omits status bar, navigation bar and cutouts when displaying WebView.
+* Added "setRequestTimeoutDuration(requestTimeoutDuration)" init config method to change request timeout duration in seconds.
+* Added setting custom network request headers support for web platform.
 
 ## 25.4.4
 * Mitigated an issue in iOS class paths.